Create Todo and FE file structure change
