+++ frontend/static/js/components/team/MemberList.js
