+++ frontend/static/js/components/team/TeamList.js
