+++ frontend/static/js/views/CreateTodoView.js
@@ -1,6 +1,6 @@
 import AbstractView from "./AbstractView.js";
-import "../components/create/CreateTodo.js";
-import "../components/shared/MainLayout.js";
+import "../components/team/CreateTodo.js";
+import "../components/team/TeamLayout.js";
 
 export default class HomeView extends AbstractView {
     constructor() {
@@ -9,7 +9,7 @@ export default class HomeView extends AbstractView {
     }
 
     async getHtml() {
-        const layout = document.createElement('main-layout');
+        const layout = document.createElement('team-layout');
         const createTodo = document.createElement('create-todo');
         layout.title = "Create Todo";
         layout.appendChild(createTodo);
